Clean up migration and seed files
CLEANUP: Consolidate to single migration and seed system

Removed Conflicts:
❌ Deleted old migration: 1750000000_core_4x_collections.go
❌ Removed duplicate old seed directory
❌ Fixed missing createCollectionFromJSON function

Unified Structure:
✅ Single migration: 1751000000_proper_4x_schema.go
✅ Single seed: cmd/seed/main.go (contains new 4X data)
✅ Clean build with no compilation errors

Benefits:
- No more schema conflicts between old/new migrations
- Single source of truth for 4X database structure
- Streamlined development workflow
- Proper separation of concerns

Both main server and seed utility now compile and run correctly.
